requirements.txt
pyTelegramBotAPI
google-generativeai
supabase
flask